Split NameToID out of get subcommand and into own utilities package. Also added a utility function to prompt user for confirmation. Credit for that code in comment above function declaration.
